+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/IconTimeBasedWeatherValueFactory.java
@@ -53,21 +53,16 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
   }
 
   @Override
-  protected List<Set<String>> getFields(TimeBasedWeatherValueData data) {
+  protected List<Set<String>> getFields(Class<?> entityClass) {
     Set<String> minParameters =
         newSet(
-            TIME,
-            DIFFUSE_IRRADIANCE,
-            DIRECT_IRRADIANCE,
-            TEMPERATURE,
-            WIND_VELOCITY_U,
-            WIND_VELOCITY_V);
+            DIFFUSE_IRRADIANCE, DIRECT_IRRADIANCE, TEMPERATURE, WIND_VELOCITY_U, WIND_VELOCITY_V);
     Set<String> allParameters =
         expandSet(
             minParameters,
             "albrad",
             "asobs",
-            "aswdifus",
+            "aswdifuS",
             "tG",
             "u10m",
             "u20m",