+++ src/main/java/edu/ie3/datamodel/io/source/EntitySource.java
@@ -6,7 +6,9 @@
 package edu.ie3.datamodel.io.source;
 
 import edu.ie3.datamodel.exceptions.FactoryException;
+import edu.ie3.datamodel.exceptions.FailedValidationException;
 import edu.ie3.datamodel.exceptions.SourceException;
+import edu.ie3.datamodel.exceptions.ValidationException;
 import edu.ie3.datamodel.io.factory.EntityFactory;
 import edu.ie3.datamodel.io.factory.SimpleEntityData;
 import edu.ie3.datamodel.io.factory.input.AssetInputEntityData;
@@ -38,6 +40,33 @@ public abstract class EntitySource {
   DataSource dataSource;
 
   // -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=
+
+  public abstract void validate() throws ValidationException;
+
+  /**
+   * Method for validating a single source.
+   *
+   * @param entityClass class to be validated
+   * @param validator used to validate
+   * @param <C> type of the class
+   */
+  protected final <C extends UniqueEntity> Try<Void, ValidationException> validate(
+      Class<? extends C> entityClass, SourceValidator<C> validator) {
+    return Try.of(() -> dataSource.getSourceFields(entityClass), SourceException.class)
+        .transformF(
+            se ->
+                (ValidationException)
+                    new FailedValidationException(
+                        "Validation for entity "
+                            + entityClass
+                            + " failed because of an error related to its source.",
+                        se))
+        .flatMap(
+            fieldsOpt ->
+                fieldsOpt
+                    .map(fields -> validator.validate(fields, entityClass))
+                    .orElse(Try.Success.empty()));
+  }
 
   protected String buildSkippingMessage(
       String entityDesc, String entityUuid, String entityId, String missingElementsString) {